requirements.txt
appdirs==1.4.3
attrs==19.1.0
black==19.3b0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
idna==2.8
python-dotenv==0.10.1
requests==2.21.0
scandir==1.10.0
six==1.12.0
toml==0.10.0
urllib3==1.24.1
